Expose method to be reusable by other tests
Apply review suggestions
